+++ Changelog.md
@@ -17,6 +17,7 @@
 
 - ### Enhancements
   - [#279](../../issues/279) - Localization of ColorGallery
+  - [#299](../../issues/299) - Quick access items should show item text as tool tip if no tooltip is set
 
 ## 4.0.3